test(ui): split markdown render-cache tests out of markdown.test.ts
Main's check-lint-core-2 job fails because markdown.test.ts reached 1011 counted lines against the 1000-line max-lines limit. PRs #141431 and #141369 passed separately but combined to cross the limit when both landed. Move the cache-separation and large-text tests into a sibling file, preserving all 115 tests and their assertions without changing production code or lint rules.
